refactor(View): 범용적인 입출력 함수와 비즈니스 로직 구현을 위한 입출력 함수를 구분한다.
UI와 밀접한 범용적인 입출력 함수를 View/UI 패키지에 구분한다.
